Push automatique :)

+++ ft_printf.h
@@ -21,7 +21,7 @@ int	ft_putchar(char c);
 int	ft_putstr(char *str);
 int	ft_putnbr(int n);
 int	ft_putnbr_u(unsigned int n);
-int	ft_puthex(unsigned int n);
+int	ft_puthex(unsigned long n);
 int	ft_puthex_upper(unsigned int n);
 int	ft_putptr(void *ptr);
 int	ft_printf(const char *format, ...);
