+++ ft_puthex.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-int	ft_puthex(unsigned int n)
+int	ft_puthex(unsigned long n)
 {
 	char	*hex;
 	int		len;
